pyproject.toml
[project]
name = "sandbox-on-ec2"
version = "1.6.0"
description = "基于EC2实例的代码执行沙箱工具，为AI Agents提供云端代码执行能力"
readme = "README.md"
license = {text = "MIT"}

requires-python = ">=3.11"

dependencies = [
    "boto3>=1.26.0",
    "strands-agents",
    "pathlib2>=2.3.0",
    "typing-extensions>=4.0.0",
    "gradio>=5.35.0",
]

[project.optional-dependencies]
dev = [
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
]
